+++ src/app/app.routes.ts
@@ -3,8 +3,6 @@ import { HederComponent } from './heder/heder.component';
 import { UsersListComponent } from './users-list/users-list.component';
 
 export const routes: Routes = [
-
-{path:'users', component: UsersListComponent},
-{path:'heder',component: HederComponent}
- 
+  { path: 'users', component: UsersListComponent },
+  { path: 'heder', component: HederComponent }
 ];
