+++ src/app/heder/heder.component.ts
@@ -11,32 +11,6 @@ const text : string = ('о компании ');
 const newPages : number [] = [5,4,3,2,1] ;
 
 
-export interface User {
-      
-    "id": number,
-    "name":string,
-    "username": string,
-    "email":string,
-    "address": {
-      "street": string,
-      "suite": string,
-      "city": string,
-      "zipcode": string,
-      "geo": {
-        "lat": string,
-        "lng": string
-      }
-    },
-    "phone": string,
-    "website": string,
-    "company": {
-      "name": string,
-      "catchPhrase": string,
-      "bs": string
-    }
-}
-
-
 @Component({
     selector: 'heder',
     templateUrl: './heder.component.html',
@@ -48,45 +22,31 @@ export interface User {
 export class HederComponent  { 
  
 readonly server = inject(HttpClient);
-heder: User [] = [];
+  heder : any  = [];
 
 constructor() {
-this.server.get<User[]>('https://jsonplaceholder.typicode.com/users').subscribe(
+this.server.get<any>('https://jsonplaceholder.typicode.com/users').subscribe(
 
    ( icloud:any) => {
         this.heder = icloud ;
-        console.log('klon:', this.heder);
    }
-    
 )
-    }
-
-    deleteHeder(id:number){
-     this.heder = this.heder.filter( 
-        (item:User) =>{
-          if(id === id){
-            return false;
-          }
-          else{
-            return true;
-          }
-        })
-    }
+}
 
 
   aboutCompany : string = text
  
   openCatalog = true ;
-  isShowPhot0 = true ; 
+  isShowPhoto = true ; 
   headeritem1 = 'главная';
   headeritem2 ='о компании';
   headeritem3 = 'каталог ';
  
-  hedertoo1 ='Каталог'
-  hedertoo2 ='Стройматериалы'
-  hedertoo3 ='Инструменты'
-  hedertoo4 ='Электрика'
-  hedertoo5='Интерьер и одежда'
+  heder1 ='Каталог'
+  heder2 ='Стройматериалы'
+  heder3 ='Инструменты'
+  heder4 ='Электрика'
+  heder5='Интерьер и одежда'
 
   newPages : number[]= newPages;
